+++ src/components/auth/RecuperarSenhaForm.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react"
 import Link from "next/link"
 import { ArrowLeft } from "lucide-react"
-import { Button } from "@/components/ui/button"
+import { Button } from "@/components/ui/button/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { RecuperarSenhaSuccess } from "@/components/auth/RecuperarSenhaSuccess"
